agents/data-agent/pyproject.toml
[project]
name = "ai-cfo-data-agent"
version = "0.1.0"
description = "Data Agent with TiDB integration for AI CFO BI Agent"
authors = [
    {name = "AI CFO Team", email = "team@ai-cfo.com"}
]
dependencies = [
    "pydantic>=2.8.0",
    "redis>=5.0.1",
    "sqlalchemy>=2.0.35",
    "pymysql>=1.1.0",
    "aiomysql>=0.2.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88
